applications/api-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
cloud-sql-python-connector[asyncpg]>=1.7.0
google-cloud-pubsub>=2.19.0
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.17.0
google-cloud-logging>=3.9.0
google-cloud-secret-manager>=2.17.0
google-cloud-aiplatform>=1.38.0
psycopg2-binary>=2.9.9
redis>=5.0.0
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-gcp-trace>=1.6.0
